Add vault-role authz, scoped roles, and admin UI
Backend: enhance authorization to accept a VaultRole in CheckRequest and include it in cache keys; evaluate built-in vault roles via builtinVaultPolicies instead of a DB stub. Scope custom-role matching by action/resource using scopedVariants and update GetCustomRolePermissionsForUser to accept resourceType/resourceID and query for scoped variants (uses ANY($2)). Add GetUserGroupIDs through recursive CTE in group repo, surface via service and new GET /api/auth/users/{id}/groups handler. Update authz logic to stop calling the unimplemented GetVaultRole and adapt tests accordingly.

Frontend: wire new admin detail routes and screens (group, role, user), update router imports and navigation to use detail pages. Migrations: consolidate system roles into a single 'General Access' role with expanded page/task/flow permissions and adjust down migration. Tests: update and add tests for builtin vault policies and vault-role behavior.
